+++ frontend/src/components/GDXForm/Fields/GDXSelect/index.tsx
@@ -9,23 +9,29 @@ import { IPickerProps, IOption } from "../../../../types";
  * @param   {IOption}   root0.fieldValue    Inital value of the select
  * @param   {Function}  root0.setFieldValue Function to handle value change
  * @param   {unknown}   root0.pickerData    All picker options
+ * @param   {string}    root0.fieldName     Field name for the select.
  * @returns {GDXSelect}                     a JSX select
  */
-export const GDXSelect: FC<IPickerProps> = ({ fieldValue, setFieldValue, pickerData }) => {
+export const GDXSelect: FC<IPickerProps> = ({
+  fieldName,
+  fieldValue,
+  setFieldValue,
+  pickerData,
+}) => {
   return (
     <>
       {!pickerData ? (
         <Skeleton variant="rectangular" width={"auto"} height={38} />
       ) : (
         <Autocomplete
-          id={pickerData?.name}
+          id={fieldName}
           options={pickerData?.definition}
           onChange={(event, option) => {
-            setFieldValue(pickerData?.name, option);
+            setFieldValue(fieldName, option);
           }}
           value={fieldValue}
           renderInput={(params: JSX.IntrinsicAttributes & TextFieldProps) => (
-            <TextField label={pickerData?.title} name={pickerData?.name} {...params} />
+            <TextField label={pickerData?.title} name={fieldName} {...params} />
           )}
           isOptionEqualToValue={(option: IOption, value: IOption) => {
             return value.value === option.value;